A lot of refactoring. All solutions are uploaded.

+++ 1-algorithmic-toolbox/week3_programming_challenges/2_maximum_value_of_the_loot/2_maximum_value_of_the_loot.cpp
@@ -1,30 +1,37 @@
-#include <iostream>
-#include <vector>
 #include <algorithm>
 #include <cassert>
+#include <iostream>
 #include <map>
+#include <vector>
 
-struct Item {
-    uint32_t value;
-    uint32_t weight;
+struct Item
+{
+    double value;
+    double weight;
 
     Item()
-      : value{ 0 }, weight{ 0 } {}
+        : value{ 0 }
+        , weight{ 0 }
+    {}
 
-    Item(uint32_t t_Value, uint32_t t_Weight):
-      value{ t_Value }, weight{ t_Weight } {}
+    Item(double t_Value, double t_Weight)
+        : value{ t_Value }
+        , weight{ t_Weight }
+    {}
 
-    float ValuePerUnitOfWeight(uint32_t t_Weight) {
-        return ((float)value / (float)weight) * (float)t_Weight;
+    double ValuePerUnitOfWeight(double t_Weight)
+    {
+        return (value / weight) * t_Weight;
     }
 };
 
-double GetOptimalValue(uint32_t capacity, std::vector<Item>& items) {
+double GetOptimalValue(uint32_t capacity, std::vector<Item>& items)
+{
     double value = 0.0;
 
     // 1. Sort the input array of items
     std::sort(items.begin(), items.end(), [](const Item& a, const Item& b) {
-        return ((float)a.value / (float)a.weight) > ((float)b.value / (float)b.weight); 
+        return (a.value / a.weight) > (b.value / b.weight);
     });
 
     // Pick the most profitable item
@@ -36,7 +43,7 @@ double GetOptimalValue(uint32_t capacity, std::vector<Item>& items) {
         }
 
         // Try to get as much as possible of the most valuable item
-        uint32_t amount = std::min(capacity, ptr->weight);
+        uint32_t amount = std::min(static_cast<double>(capacity), ptr->weight);
 
         value += ptr->ValuePerUnitOfWeight(amount);
         capacity -= amount;
@@ -50,7 +57,8 @@ double GetOptimalValue(uint32_t capacity, std::vector<Item>& items) {
     return value;
 }
 
-int main(int argc, char* argv[]) {
+int main(int argc, char* argv[])
+{
     uint32_t n;
     uint32_t capacity;
     std::cin >> n >> capacity;
